Create AddAllyStatusEffect.cs
Created a new event that allows you to directly apply a status effect to the closest ally player within a specified range.  If "AFfectController" is true it is capable of targeting the caster if no other player is nearby, otherwise it will do nothing if another player is not in range.